Added exception handling

+++ src/main/java/com/epam/druzhinin/repositories/GoodsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GoodsRepository extends JpaRepository<GoodsEntity, Integer> {
+public interface GoodsRepository extends JpaRepository<GoodsEntity, Long> {
 }
